Make the squares generate in flex rows
Initially I wanted to set the width and height of the squares using
square.setAttribute, but it didn't work. Apparently, the style of the
grid is more specific and overwrites the style of the squares.

Instead I decided to create rows arranged in column. They would take
up all the space in the grid (flex: 1), then squares would be added
to the rows which would too distribute evenly all the available space
in the row.

Thus I don't need a setSize or getSize anymore, flexbox does it for me.